Refactor for PersistenceEntityQuickFix.java
Moved generic quickFixes under codeActions from DependencyInjection
refactored PersistenceEntityQuickFix.java to utilize RemoveFinalModifierQuickFix.java
