+++ jakarta.jdt/org.eclipse.lsp4jakarta.jdt.core/src/main/java/org/eclipse/lsp4jakarta/jdt/codeAction/proposal/quickfix/RemoveModifierConflictQuickFix.java
@@ -107,13 +107,13 @@ public class RemoveModifierConflictQuickFix implements IJavaCodeActionParticipan
         // API
         ASTNode coveredNode = context.getCoveredNode().getParent();
         String type = "";
-        if (diagnostic.getData().equals(String.valueOf(IJavaElement.LOCAL_VARIABLE))){
+        if (diagnostic.getData().toString().equals(String.valueOf(IJavaElement.LOCAL_VARIABLE))){
             type = "variable";
         } else if (diagnostic.getData().toString().equals(String.valueOf(IJavaElement.FIELD))) {
             type = "field";
-        } else if (diagnostic.getData().equals(String.valueOf(IJavaElement.METHOD))) {
+        } else if (diagnostic.getData().toString().equals(String.valueOf(IJavaElement.METHOD))) {
             type = "method";
-        } else if (diagnostic.getData().equals(String.valueOf(IJavaElement.CLASS_FILE))) {
+        } else if (diagnostic.getData().toString().equals(String.valueOf(IJavaElement.CLASS_FILE))) {
             type = "class";
         }
